Demo 4: AS: Add /oauth2/authorize endpoint
If there's no user session, the browser will be redirected to the login page (via the [Authorize] attribute).
When a user is authorized, the endpoint will create/update a record in the Grants table representing the OAuth2 grant.
It will also create a new record in the AuthorizationCodes table, which is needed for the upcoming /oauth2/token endpoint.
First-party client support only in this commit.

Create migration using this command:
dotnet ef migrations add CreateOAuth2Schema